# app7/swagger.yaml
openapi: 3.0.3
info:
  title: API de Previsão de Preços de Imóveis com Banco de Dados
  description: |
    Uma API RESTful para previsão de preços de imóveis com persistência em banco de dados.
    
    Funcionalidades:
    * Treinamento de modelos de machine learning
    * Previsão de preços
    * Armazenamento de histórico de treinamentos e previsões
    * Consulta de estatísticas
  
  version: "7.0.0"
  contact:
    name: Suporte
    email: suporte@exemplo.com
servers:
  - url: http://localhost:8000
    description: Servidor de desenvolvimento local
...
paths:
...
  /prever:
    post:
      tags:
        - Previsão
      summary: Prever preço
      description: Realiza previsão de preço com base nas características do imóvel
      operationId: predictPrice
      requestBody:
        description: Características do imóvel
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrevisaoInput'
        required: true
      responses:
        '200':
          description: Previsão bem-sucedida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrevisaoOutput'
        '400':
          description: Entrada inválida ou modelo não treinado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /treinar:
    post:
      tags:
        - Treinamento
      summary: Treinar modelo
      description: Treina o modelo com dados fornecidos
      operationId: trainModel
      requestBody:
        description: Dados para treinamento do modelo
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreinamentoInput'
        required: true
      responses:
        '200':
          description: Treinamento bem-sucedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                    example: "f8d2e580-1b3a-4c1a-8c8e-5a5f2a7f9e6c"
                  mensagem:
                    type: string
                    example: "Modelo treinado e salvo com sucesso"
                  num_amostras:
                    type: integer
                    example: 10
                  r2_score:
                    type: number
                    format: float
                    example: 0.95
                  rmse:
                    type: number
                    format: float
                    example: 15000.5
                  mae:
                    type: number
                    format: float
                    example: 12500.2
                  training_time_seconds:
                    type: number
                    format: float
                    example: 0.156
                  features_utilizadas:
                    type: array
                    items:
                      type: string
                    example: ["area", "quartos", "banheiros", "idade_imovel"]
        '400':
          description: Erro de treinamento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /previsoes:
    get:
      tags:
        - Histórico
      summary: Histórico de previsões
      description: Retorna o histórico de previsões realizadas
      operationId: getPredictionHistory
      parameters:
        - name: limit
          in: query
          description: Número máximo de registros a retornar
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          description: Índice a partir do qual retornar registros
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Operação bem-sucedida
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistoricoPrevisao'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    PrevisaoInput:
      type: object
      required:
        - area
      properties:
        area:
          type: number
          format: float
          description: Área do imóvel em metros quadrados
          minimum: 0
          exclusiveMinimum: true
          example: 100.5
        quartos:
          type: integer
          description: Número de quartos
          minimum: 0
          example: 2
        banheiros:
          type: integer
          description: Número de banheiros
          minimum: 0
          example: 1
        idade_imovel:
          type: integer
          description: Idade do imóvel em anos
          minimum: 0
          example: 5
    PrevisaoOutput:
      type: object
      properties:
        request_id:
          type: string
          description: Identificador único da requisição
          example: "f8d2e580-1b3a-4c1a-8c8e-5a5f2a7f9e6c"
        timestamp:
          type: string
          format: date-time
          description: Data e hora da previsão
          example: "2023-05-01T12:00:00"
        area:
          type: number
          format: float
          description: Área do imóvel
          example: 100.5
        quartos:
          type: integer
          nullable: true
          description: Número de quartos
          example: 2
        banheiros:
          type: integer
          nullable: true
          description: Número de banheiros
          example: 1
        idade_imovel:
          type: integer
          nullable: true
          description: Idade do imóvel em anos
          example: 5
        preco_previsto:
          type: number
          format: float
          description: Preço previsto
          example: 350000.0
        faixa_confianca:
          type: array
          description: Faixa de confiança para a previsão
          minItems: 2
          maxItems: 2
          items:
            type: number
            format: float
          example: [315000.0, 385000.0]
    TreinamentoInput:
      type: object
      required:
        - features
        - precos
      properties:
        features:
          type: array
          description: Lista de características dos imóveis
          items:
            type: object
            required:
              - area
            properties:
              area:
                type: number
                format: float
                description: Área do imóvel em metros quadrados
                example: 100.5
              quartos:
                type: integer
                nullable: true
                description: Número de quartos
                example: 2
              banheiros:
                type: integer
                nullable: true
                description: Número de banheiros
                example: 1
              idade_imovel:
                type: integer
                nullable: true
                description: Idade do imóvel em anos
                example: 5
          example:
            - area: 100.5
              quartos: 2
              banheiros: 1
              idade_imovel: 5
            - area: 150.0
              quartos: 3
              banheiros: 2
              idade_imovel: 3
        precos:
          type: array
          description: Lista de preços correspondentes aos imóveis
          items:
            type: number
            format: float
          example: [350000.0, 450000.0]
...
    HistoricoPrevisao:
      type: object
      properties:
        request_id:
          type: string
          description: Identificador único da requisição
          example: "f8d2e580-1b3a-4c1a-8c8e-5a5f2a7f9e6c"
        timestamp:
          type: string
          format: date-time
          description: Data e hora da previsão
          example: "2023-05-01T12:00:00"
        input:
          type: object
          description: Características do imóvel
          properties:
            area:
              type: number
              format: float
              example: 100.5
            quartos:
              type: integer
              nullable: true
              example: 2
            banheiros:
              type: integer
              nullable: true
              example: 1
            idade_imovel:
              type: integer
              nullable: true
              example: 5
        preco_previsto:
          type: number
          format: float
          description: Preço previsto
          example: 350000.0
        faixa_confianca:
          type: array
          description: Faixa de confiança para a previsão
          minItems: 2
          maxItems: 2
          items:
            type: number
            format: float
          example: [315000.0, 385000.0]
...
    ErrorResponse:
      type: object
      properties:
        request_id:
          type: string
          description: Identificador único da requisição
          example: "f8d2e580-1b3a-4c1a-8c8e-5a5f2a7f9e6c"
        timestamp:
          type: string
          format: date-time
          description: Data e hora do erro
          example: "2023-05-01T12:00:00"
        error:
          type: string
          description: Tipo de erro
          example: "Validation Error"
        detail:
          type: string
          description: Detalhes do erro
          example: "Campo 'area': valor deve ser maior que 0"
        path:
          type: string
          description: Caminho da requisição
          example: "/prever"
        http_status:
          type: integer
          description: Código HTTP do erro
          example: 422 